[web][audio] Refactor large web file
