+++ packages/expo-audio/src/AudioModule.web.ts
@@ -1,42 +1,10 @@
-import { Asset } from 'expo-asset';
 import { PermissionResponse, PermissionStatus } from 'expo-modules-core';
 
-import {
-  AudioMetadata,
-  AudioMode,
-  AudioPlayerOptions,
-  AudioSource,
-  AudioStatus,
-  PitchCorrectionQuality,
-  RecorderState,
-  RecordingInput,
-  RecordingOptions,
-  RecordingOptionsWeb,
-  RecordingStartOptions,
-} from './Audio.types';
-import { AudioLockScreenOptions } from './AudioConstants';
-import {
-  AUDIO_SAMPLE_UPDATE,
-  PLAYBACK_STATUS_UPDATE,
-  RECORDING_STATUS_UPDATE,
-} from './AudioEventKeys';
-import { AudioPlayer, AudioEvents, RecordingEvents, AudioRecorder } from './AudioModule.types';
-import { mediaSessionController } from './MediaSessionController.web';
-import { RecordingPresets } from './RecordingConstants';
+import { AudioMode } from './Audio.types';
+import { getUserMedia } from './AudioUtils.web';
 
-const nextId = (() => {
-  let id = 0;
-  return () => id++;
-})();
-
-let audioContext: AudioContext | null = null;
-
-function getAudioContext(): AudioContext {
-  if (!audioContext) {
-    audioContext = new AudioContext();
-  }
-  return audioContext;
-}
+export { AudioPlayerWeb } from './AudioPlayer.web';
+export { AudioRecorderWeb } from './AudioRecorder.web';
 
 async function getPermissionWithQueryAsync(
   name: PermissionNameWithAdditionalValues
@@ -56,711 +24,6 @@ async function getPermissionWithQueryAsync(
   } catch {
     // Firefox - TypeError: 'microphone' (value of 'name' member of PermissionDescriptor) is not a valid value for enumeration PermissionName.
     return PermissionStatus.UNDETERMINED;
-  }
-}
-
-function getUserMedia(constraints: MediaStreamConstraints): Promise<MediaStream> {
-  if (navigator.mediaDevices && navigator.mediaDevices.getUserMedia) {
-    return navigator.mediaDevices.getUserMedia(constraints);
-  }
-
-  // Some browsers partially implement mediaDevices. We can't just assign an object
-  // with getUserMedia as it would overwrite existing properties.
-  // Here, we will just add the getUserMedia property if it's missing.
-
-  // First get ahold of the legacy getUserMedia, if present
-  const getUserMedia =
-    // TODO: this method is deprecated, migrate to https://developer.mozilla.org/en-US/docs/Web/API/MediaDevices/getUserMedia
-    navigator.getUserMedia ||
-    navigator.webkitGetUserMedia ||
-    navigator.mozGetUserMedia ||
-    function () {
-      const error: any = new Error('Permission unimplemented');
-      error.code = 0;
-      error.name = 'NotAllowedError';
-      throw error;
-    };
-
-  return new Promise((resolve, reject) => {
-    // TODO(@kitten): The types indicates that this is incorrect.
-    // Please check whether this is correct!
-    // @ts-expect-error: The `successCallback` doesn't match a `resolve` function
-    getUserMedia.call(navigator, constraints, resolve, reject);
-  });
-}
-
-function getStatusFromMedia(media: HTMLMediaElement, id: number): AudioStatus {
-  const isPlaying = !!(
-    media.currentTime > 0 &&
-    !media.paused &&
-    !media.ended &&
-    media.readyState > 2
-  );
-
-  const status: AudioStatus = {
-    id,
-    isLoaded: true,
-    duration: media.duration,
-    currentTime: media.currentTime,
-    playbackState: '',
-    timeControlStatus: isPlaying ? 'playing' : 'paused',
-    reasonForWaitingToPlay: '',
-    playing: isPlaying,
-    didJustFinish: media.ended,
-    isBuffering: false,
-    playbackRate: media.playbackRate,
-    shouldCorrectPitch: true,
-    mute: media.muted,
-    loop: media.loop,
-  };
-
-  return status;
-}
-
-export class AudioPlayerWeb
-  extends globalThis.expo.SharedObject<AudioEvents>
-  implements AudioPlayer
-{
-  constructor(source: AudioSource, options: AudioPlayerOptions = {}) {
-    super();
-    const { updateInterval = 500, crossOrigin } = options;
-    this.src = source;
-    this.interval = Math.max(updateInterval, 1);
-    this.crossOrigin = crossOrigin;
-    this.media = this._createMediaElement();
-  }
-
-  id: number = nextId();
-  isAudioSamplingSupported = false;
-  isBuffering = false;
-  shouldCorrectPitch = false;
-
-  private src: AudioSource = null;
-  private media: HTMLAudioElement;
-  private interval = 500;
-  private isPlaying = false;
-  private loaded = false;
-  private crossOrigin?: 'anonymous' | 'use-credentials';
-  private analyser: AnalyserNode | null = null;
-  private sourceNode: MediaElementAudioSourceNode | null = null;
-  private samplingFrameId: number | null = null;
-  private samplingEnabled = false;
-
-  get playing(): boolean {
-    return this.isPlaying;
-  }
-
-  get muted(): boolean {
-    return this.media.muted;
-  }
-
-  set muted(value: boolean) {
-    this.media.muted = value;
-  }
-
-  get loop(): boolean {
-    return this.media.loop;
-  }
-
-  set loop(value: boolean) {
-    this.media.loop = value;
-  }
-
-  get duration(): number {
-    return this.media.duration;
-  }
-
-  get currentTime(): number {
-    return this.media.currentTime;
-  }
-
-  get paused(): boolean {
-    return this.media.paused;
-  }
-
-  get isLoaded(): boolean {
-    return this.loaded;
-  }
-
-  get playbackRate(): number {
-    return this.media.playbackRate;
-  }
-
-  set playbackRate(value: number) {
-    this.media.playbackRate = value;
-  }
-
-  get volume(): number {
-    return this.media.volume;
-  }
-
-  set volume(value: number) {
-    this.media.volume = value;
-  }
-
-  get currentStatus(): AudioStatus {
-    return getStatusFromMedia(this.media, this.id);
-  }
-
-  play(): void {
-    this.media.play();
-    this.isPlaying = true;
-  }
-
-  pause(): void {
-    this.media.pause();
-    this.isPlaying = false;
-  }
-
-  replace(source: AudioSource): void {
-    const wasPlaying = this.isPlaying;
-    const wasSampling = this.samplingEnabled;
-    const mediaSessionState = mediaSessionController.getActiveState(this);
-
-    // we need to remove the current media element and create a new one
-    this.remove();
-
-    this.src = source;
-    this.isPlaying = false;
-    this.loaded = false;
-    this.media = this._createMediaElement();
-
-    if (wasSampling) {
-      this.setAudioSamplingEnabled(true);
-    }
-
-    if (mediaSessionState) {
-      mediaSessionController.setActivePlayer(
-        this,
-        mediaSessionState.metadata ?? undefined,
-        mediaSessionState.options ?? undefined
-      );
-    }
-
-    // Resume playback if it was playing before
-    if (wasPlaying) {
-      this.play();
-    }
-  }
-
-  async seekTo(
-    seconds: number,
-    toleranceMillisBefore?: number,
-    toleranceMillisAfter?: number
-  ): Promise<void> {
-    this.media.currentTime = seconds;
-  }
-
-  setAudioSamplingEnabled(enabled: boolean): void {
-    if (enabled) {
-      if (!this.media.crossOrigin && this._isCrossOrigin()) {
-        this.isAudioSamplingSupported = false;
-        return;
-      }
-
-      if (this.analyser) {
-        return;
-      }
-
-      const ctx = getAudioContext();
-      if (ctx.state === 'suspended') {
-        ctx.resume();
-      }
-
-      if (!this.sourceNode) {
-        this.sourceNode = ctx.createMediaElementSource(this.media);
-      }
-
-      this.analyser = ctx.createAnalyser();
-      this.analyser.fftSize = 2048;
-
-      this.sourceNode.disconnect();
-      this.sourceNode.connect(this.analyser);
-      this.analyser.connect(ctx.destination);
-
-      const buffer = new Float32Array(this.analyser.frequencyBinCount);
-
-      const sampleLoop = () => {
-        if (!this.analyser) {
-          return;
-        }
-        if (this.isPlaying) {
-          this.analyser.getFloatTimeDomainData(buffer);
-          this.emit(AUDIO_SAMPLE_UPDATE, {
-            channels: [{ frames: Array.from(buffer) }],
-            timestamp: this.media.currentTime,
-          });
-        }
-        this.samplingFrameId = requestAnimationFrame(sampleLoop);
-      };
-      this.samplingFrameId = requestAnimationFrame(sampleLoop);
-
-      this.samplingEnabled = true;
-      this.isAudioSamplingSupported = true;
-    } else {
-      if (this.samplingFrameId != null) {
-        cancelAnimationFrame(this.samplingFrameId);
-        this.samplingFrameId = null;
-      }
-
-      if (this.analyser) {
-        this.analyser.disconnect();
-        this.analyser = null;
-      }
-
-      if (this.sourceNode) {
-        const ctx = getAudioContext();
-        this.sourceNode.disconnect();
-        this.sourceNode.connect(ctx.destination);
-      }
-
-      this.samplingEnabled = false;
-    }
-  }
-
-  setPlaybackRate(second: number, pitchCorrectionQuality?: PitchCorrectionQuality): void {
-    this.media.playbackRate = second;
-    this.shouldCorrectPitch = pitchCorrectionQuality === 'high';
-    this.media.preservesPitch = this.shouldCorrectPitch;
-    mediaSessionController.updatePositionState(this);
-  }
-
-  remove(): void {
-    mediaSessionController.clear(this);
-
-    if (this.samplingFrameId != null) {
-      cancelAnimationFrame(this.samplingFrameId);
-      this.samplingFrameId = null;
-    }
-
-    if (this.analyser) {
-      this.analyser.disconnect();
-      this.analyser = null;
-    }
-
-    if (this.sourceNode) {
-      this.sourceNode.disconnect();
-      this.sourceNode = null;
-    }
-
-    this.samplingEnabled = false;
-    this.media.pause();
-    this.media.removeAttribute('src');
-    this.media.load();
-    getStatusFromMedia(this.media, this.id);
-  }
-
-  setActiveForLockScreen(
-    active: boolean,
-    metadata?: AudioMetadata,
-    options?: AudioLockScreenOptions
-  ): void {
-    if (active) {
-      mediaSessionController.setActivePlayer(this, metadata, options);
-    } else {
-      mediaSessionController.clear(this);
-    }
-  }
-
-  updateLockScreenMetadata(metadata: AudioMetadata): void {
-    mediaSessionController.updateMetadata(this, metadata);
-  }
-
-  clearLockScreenControls(): void {
-    mediaSessionController.clear(this);
-  }
-
-  _isCrossOrigin(): boolean {
-    try {
-      return new URL(this.media.src).origin !== window.location.origin;
-    } catch {
-      return false;
-    }
-  }
-
-  _createMediaElement(): HTMLAudioElement {
-    const newSource = getSourceUri(this.src);
-    const media = new Audio(newSource);
-    if (this.crossOrigin !== undefined) {
-      media.crossOrigin = this.crossOrigin;
-    }
-
-    let lastEmitTime = 0;
-    const intervalSec = this.interval / 1000;
-
-    // Throttled status updates based on interval
-    media.ontimeupdate = () => {
-      const now = media.currentTime;
-      // Handle backwards time (loop/seek)
-      if (now < lastEmitTime) {
-        lastEmitTime = now;
-      }
-      if (now - lastEmitTime >= intervalSec) {
-        lastEmitTime = now;
-        this.emit(PLAYBACK_STATUS_UPDATE, getStatusFromMedia(media, this.id));
-        mediaSessionController.updatePositionState(this);
-      }
-    };
-
-    media.onplay = () => {
-      this.isPlaying = true;
-      lastEmitTime = media.currentTime;
-      this.emit(PLAYBACK_STATUS_UPDATE, {
-        ...getStatusFromMedia(media, this.id),
-        playing: this.isPlaying,
-      });
-      mediaSessionController.updatePlaybackState(this);
-      mediaSessionController.updatePositionState(this);
-    };
-
-    media.onpause = () => {
-      this.isPlaying = false;
-      lastEmitTime = media.currentTime;
-      this.emit(PLAYBACK_STATUS_UPDATE, {
-        ...getStatusFromMedia(media, this.id),
-        playing: this.isPlaying,
-      });
-      mediaSessionController.updatePlaybackState(this);
-      mediaSessionController.updatePositionState(this);
-    };
-
-    media.onseeked = () => {
-      lastEmitTime = media.currentTime;
-      this.emit(PLAYBACK_STATUS_UPDATE, getStatusFromMedia(media, this.id));
-      mediaSessionController.updatePositionState(this);
-    };
-
-    media.onended = () => {
-      lastEmitTime = 0;
-      mediaSessionController.updatePlaybackState(this);
-    };
-
-    media.onloadeddata = () => {
-      this.loaded = true;
-      lastEmitTime = media.currentTime;
-      this.emit(PLAYBACK_STATUS_UPDATE, {
-        ...getStatusFromMedia(media, this.id),
-        isLoaded: this.loaded,
-      });
-      mediaSessionController.updatePositionState(this);
-    };
-
-    return media;
-  }
-}
-
-function getSourceUri(source: AudioSource): string | undefined {
-  if (typeof source === 'string') {
-    return source;
-  }
-  if (typeof source === 'number') {
-    const asset = Asset.fromModule(source);
-    return asset.uri;
-  }
-  if (typeof source?.assetId === 'number' && !source?.uri) {
-    const asset = Asset.fromModule(source.assetId);
-    return asset.uri;
-  }
-
-  return source?.uri ?? undefined;
-}
-
-export class AudioRecorderWeb
-  extends globalThis.expo.SharedObject<RecordingEvents>
-  implements AudioRecorder
-{
-  constructor(options: Partial<RecordingOptions>) {
-    super();
-    this.options = options;
-  }
-
-  async setup() {
-    this.mediaRecorder = await this.createMediaRecorder(this.options);
-  }
-
-  id = nextId();
-  currentTime = 0;
-  uri: string | null = null;
-
-  private options: Partial<RecordingOptions>;
-  private mediaRecorder: MediaRecorder | null = null;
-  private mediaRecorderUptimeOfLastStartResume = 0;
-  private mediaRecorderIsRecording = false;
-  private timeoutIds: number[] = [];
-  private cachedInputs: RecordingInput[] = [];
-  private selectedDeviceId: string | null = null;
-  private stream: MediaStream | null = null;
-  private handleDeviceChange: (() => void) | null = null;
-  private analyser: AnalyserNode | null = null;
-  private analyserBuffer: Float32Array<ArrayBuffer> | null = null;
-  private analyserSource: MediaStreamAudioSourceNode | null = null;
-  private meteringEnabled = false;
-
-  get isRecording(): boolean {
-    return this.mediaRecorder?.state === 'recording';
-  }
-
-  record(options?: RecordingStartOptions): void {
-    if (this.mediaRecorder === null) {
-      throw new Error(
-        'Cannot start an audio recording without initializing a MediaRecorder. Run prepareToRecordAsync() before attempting to start an audio recording.'
-      );
-    }
-
-    // Clear any existing timeouts
-    this.clearTimeouts();
-
-    // Note: atTime is not supported on Web (no native equivalent), so we ignore it entirely
-    // Only forDuration is implemented using setTimeout
-    const { forDuration } = options || {};
-
-    this.startActualRecording();
-
-    if (forDuration !== undefined) {
-      this.timeoutIds.push(
-        setTimeout(() => {
-          this.stop();
-        }, forDuration * 1000)
-      );
-    }
-  }
-
-  private startActualRecording(): void {
-    if (this.mediaRecorder?.state === 'paused') {
-      this.mediaRecorder.resume();
-    } else {
-      this.mediaRecorder?.start();
-    }
-  }
-
-  getAvailableInputs(): RecordingInput[] {
-    return this.cachedInputs;
-  }
-
-  getCurrentInput(): Promise<RecordingInput> {
-    const deviceId =
-      this.selectedDeviceId ?? this.stream?.getAudioTracks()[0]?.getSettings().deviceId;
-    const matched = this.cachedInputs.find((input) => input.uid === deviceId);
-    if (matched) {
-      return Promise.resolve(matched);
-    }
-    return Promise.resolve(
-      this.cachedInputs[0] ?? {
-        type: 'Default',
-        name: 'Default',
-        uid: 'Default',
-      }
-    );
-  }
-
-  async prepareToRecordAsync(): Promise<void> {
-    return this.setup();
-  }
-
-  getStatus(): RecorderState {
-    const status: RecorderState = {
-      canRecord:
-        this.mediaRecorder?.state === 'recording' || this.mediaRecorder?.state === 'inactive',
-      isRecording: this.mediaRecorder?.state === 'recording',
-      durationMillis: this.getAudioRecorderDurationMillis(),
-      mediaServicesDidReset: false,
-      url: this.uri,
-    };
-    if (this.meteringEnabled && this.mediaRecorderIsRecording) {
-      status.metering = this.getMeteringLevel();
-    }
-    return status;
-  }
-
-  pause(): void {
-    if (this.mediaRecorder === null) {
-      throw new Error(
-        'Cannot start an audio recording without initializing a MediaRecorder. Run prepareToRecordAsync() before attempting to start an audio recording.'
-      );
-    }
-
-    this.mediaRecorder?.pause();
-  }
-
-  recordForDuration(seconds: number): void {
-    this.record({ forDuration: seconds });
-  }
-
-  setInput(input: string): void {
-    if (!this.cachedInputs.some((cached) => cached.uid === input)) {
-      throw new Error(`No audio input device found for uid: ${input}`);
-    }
-    this.selectedDeviceId = input;
-  }
-
-  startRecordingAtTime(seconds: number): void {
-    this.record({ atTime: seconds });
-  }
-
-  async stop(): Promise<void> {
-    if (this.mediaRecorder === null) {
-      throw new Error(
-        'Cannot start an audio recording without initializing a MediaRecorder. Run prepareToRecordAsync() before attempting to start an audio recording.'
-      );
-    }
-
-    const dataPromise = new Promise<Blob>((resolve) =>
-      this.mediaRecorder?.addEventListener('dataavailable', (e) => resolve(e.data))
-    );
-
-    this.mediaRecorder?.stop();
-    this.mediaRecorder = null;
-
-    const data = await dataPromise;
-    const url = URL.createObjectURL(data);
-
-    this.uri = url;
-
-    this.emit(RECORDING_STATUS_UPDATE, {
-      id: this.id,
-      isFinished: true,
-      hasError: false,
-      error: null,
-      url,
-    });
-  }
-
-  clearTimeouts() {
-    this.timeoutIds.forEach((id) => clearTimeout(id));
-  }
-
-  private async createMediaRecorder(
-    options: Partial<RecordingOptions> & Partial<RecordingOptionsWeb>
-  ): Promise<MediaRecorder> {
-    if (typeof navigator !== 'undefined' && !navigator.mediaDevices) {
-      throw new Error('No media devices available');
-    }
-
-    this.mediaRecorderUptimeOfLastStartResume = 0;
-    this.currentTime = 0;
-
-    const audioConstraints = this.selectedDeviceId
-      ? { deviceId: { exact: this.selectedDeviceId } }
-      : true;
-    const stream = await getUserMedia({ audio: audioConstraints });
-    this.stream = stream;
-
-    await this.updateCachedInputs();
-
-    this.handleDeviceChange = () => {
-      this.updateCachedInputs();
-    };
-    navigator.mediaDevices.addEventListener('devicechange', this.handleDeviceChange);
-
-    if (options.isMeteringEnabled) {
-      const ctx = getAudioContext();
-      if (ctx.state === 'suspended') {
-        ctx.resume();
-      }
-      this.analyserSource = ctx.createMediaStreamSource(stream);
-      this.analyser = ctx.createAnalyser();
-      this.analyser.fftSize = 2048;
-      this.analyserSource.connect(this.analyser);
-      this.analyserBuffer = new Float32Array(this.analyser.frequencyBinCount);
-      this.meteringEnabled = true;
-    }
-
-    const defaults = RecordingPresets.HIGH_QUALITY.web;
-    const mediaRecorderOptions: MediaRecorderOptions = {};
-
-    const mimeType = options.mimeType ?? defaults.mimeType;
-    if (mimeType && MediaRecorder.isTypeSupported(mimeType)) {
-      mediaRecorderOptions.mimeType = mimeType;
-    }
-
-    if (options.bitsPerSecond) {
-      mediaRecorderOptions.bitsPerSecond = options.bitsPerSecond;
-    } else if (options.bitRate) {
-      mediaRecorderOptions.audioBitsPerSecond = options.bitRate;
-    } else {
-      mediaRecorderOptions.bitsPerSecond = defaults.bitsPerSecond;
-    }
-
-    const mediaRecorder = new MediaRecorder(stream, mediaRecorderOptions);
-
-    mediaRecorder.addEventListener('pause', () => {
-      this.currentTime = this.getAudioRecorderDurationMillis();
-      this.mediaRecorderIsRecording = false;
-    });
-
-    mediaRecorder.addEventListener('resume', () => {
-      this.mediaRecorderUptimeOfLastStartResume = Date.now();
-      this.mediaRecorderIsRecording = true;
-    });
-
-    mediaRecorder.addEventListener('start', () => {
-      this.mediaRecorderUptimeOfLastStartResume = Date.now();
-      this.currentTime = 0;
-      this.mediaRecorderIsRecording = true;
-    });
-
-    mediaRecorder?.addEventListener('stop', () => {
-      this.currentTime = 0;
-      this.mediaRecorderIsRecording = false;
-      this.stream = null;
-
-      if (this.analyserSource) {
-        this.analyserSource.disconnect();
-        this.analyserSource = null;
-      }
-      if (this.analyser) {
-        this.analyser.disconnect();
-        this.analyser = null;
-        this.analyserBuffer = null;
-      }
-
-      if (this.handleDeviceChange) {
-        navigator.mediaDevices?.removeEventListener('devicechange', this.handleDeviceChange);
-        this.handleDeviceChange = null;
-      }
-
-      // Clears recording icon in Chrome tab
-      stream.getTracks().forEach((track) => track.stop());
-    });
-
-    return mediaRecorder;
-  }
-
-  private async updateCachedInputs() {
-    const devices = await navigator.mediaDevices.enumerateDevices();
-    this.cachedInputs = devices
-      .filter((device) => device.kind === 'audioinput')
-      .map((device) => ({
-        uid: device.deviceId,
-        name: device.label || 'Unknown Device',
-        type: device.deviceId === 'default' ? 'Default' : 'Unknown',
-      }));
-  }
-
-  // Compute the metering level in dBFS using the same RMS-to-decibel formula as
-  // native (20 * log10(rms)). -160 represents silence / no signal.
-  private getMeteringLevel(): number {
-    if (!this.analyser || !this.analyserBuffer) {
-      return -160;
-    }
-    this.analyser.getFloatTimeDomainData(this.analyserBuffer);
-    let sumSquares = 0;
-    for (let i = 0; i < this.analyserBuffer.length; i++) {
-      sumSquares += this.analyserBuffer[i] * this.analyserBuffer[i];
-    }
-    const rms = Math.sqrt(sumSquares / this.analyserBuffer.length);
-    if (rms === 0) {
-      return -160;
-    }
-    return 20 * Math.log10(rms);
-  }
-
-  private getAudioRecorderDurationMillis() {
-    let duration = this.currentTime;
-    if (this.mediaRecorderIsRecording && this.mediaRecorderUptimeOfLastStartResume > 0) {
-      duration += Date.now() - this.mediaRecorderUptimeOfLastStartResume;
-    }
-    return duration;
   }
 }
